% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/best_worst_month.R
\name{best_worst_month}
\alias{best_worst_month}
\title{Find the best-performing and the worst-performing months}
\usage{
best_worst_month(x)
}
\arguments{
\item{x}{A data frame that contains data for individual instruments.}
}
\value{
A list with the best month and the worst month, as well as their
respective p&l's.
}
\description{
This function takes in a data set and returns a list with the best and the
worst month and their respective pnl's.
}

