% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/best_worst_three.R
\name{best_worst_three}
\alias{best_worst_three}
\title{Find the three best-performing and worst-performing commodities}
\usage{
best_worst_three(x)
}
\arguments{
\item{x}{A data frame that contains data for individual commodities.}
}
\value{
output A list of the best three and worst three performers with their
commodity names, ids and respective pnls
}
\description{
This function takes in a data set and returns the best three and worst three
commodity names and their respective pnls. All these data will be presented
as a formatted table.
}

