% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_pnl.R
\name{calc_pnl}
\alias{calc_pnl}
\title{Calculate cumulative pnl, mean pnl, dollar sharpe ratio}
\usage{
calc_pnl(x, trade.freq = 7)
}
\arguments{
\item{x}{A data frame that contains data for individual commodities.}

\item{trade.freq}{The trading frequency of the portfolio, numeric}
}
\value{
a list ith cumulative pnl, mean pnl, annualized volatility of pnl,
dollar sharpe ratio and drawdown
}
\description{
This function takes in a data frame and calculates the cumulative P&L, daily
P&L, annualized P&L, P&L volatility and dollar sharpe ratio as well as three
biggest drawdowns. There are also average daily return rate and the standard
deviation of daily return rates.
}

