% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_instrument_optGroups.R
\name{create_instrument_optGroups}
\alias{create_instrument_optGroups}
\title{Create opt group list for selectizeInput for instruments}
\usage{
create_instrument_optGroups(sector, id)
}
\arguments{
\item{sector}{sector column}

\item{id}{id column}
}
\value{
a list of groups for sector/ id 
        as required by the selectizeInput optgroup functionality
}
\description{
This function accepts the sector and id columns
of the data set and create a list of groups as required
by the selectizeInput optgroup functionality
}
