% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_strategy_optGroups.R
\name{create_strategy_optGroups}
\alias{create_strategy_optGroups}
\title{Create opt group list for selectizeInput for strategies}
\usage{
create_strategy_optGroups(strategy, substrategy)
}
\arguments{
\item{strategy}{strategy column}

\item{substrategy}{substrategy column}
}
\value{
a list of groups for strategy/ substrategy 
        as required by the selectizeInput optgroup functionality
}
\description{
This function accepts the strategy and substrategy columns
of the data set and create a list of groups as required
by the selectizeInput optgroup functionality
}
\details{
We assume that substrategies pertaining to a given strategy 
uses the name of the strategy followed by a period and then
the substrategy number. For example, substrategy 2 for strategy
1 must be named "Strategy 1.2".
}
