% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cumpnl_plot.R
\name{cumpnl_plot}
\alias{cumpnl_plot}
\title{Draw an interactive line plot for cumulative P&L}
\usage{
cumpnl_plot(x)
}
\arguments{
\item{x}{A data frame that the function takes in.}
}
\value{
A plot. Red bars indicate loss and green bars indicate profit.
}
\description{
This function takes a data frame and returns an interactive plot for the
cumulative P&L of the portfolio data. The plot is zoomable and specific data can be shown once the
cursor is placed on the graph.
}

