% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drawdown.R
\name{drawdown}
\alias{drawdown}
\title{The Three Biggest Drawdowns in the portfolio}
\usage{
drawdown(x)
}
\arguments{
\item{x}{A data frame with date and return columns}
}
\value{
A data frame that contains the starting date, end date and values of
the three biggest drawdowns.
}
\description{
Show the top 3 drawdowns including start and end dates, as well as
decrease in returns during the drawdown period. All the information will be
returned in a table, with all the numbers properly formatted.
}
\details{
If the data set is not big enough that there are fewer drawdowns than
required by the user, the function will throw NA's into the table so that the
table will still contain as many rows as the user demands.
}

