% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactive_plot.R
\name{interactive_plot}
\alias{interactive_plot}
\title{Interactive plot with dygraph}
\usage{
interactive_plot(x, type)
}
\arguments{
\item{x}{A data frame that the function takes in.}

\item{type}{The value depicted by the plot, choices are \code{nmv}, \code{gmv},
\code{contract}, \code{pnl} and \code{ret}.}
}
\value{
A plot. Red bars indicate loss and green bars indicate profits.
}
\description{
This function takes a data frame and returns an interactive plot for
the portfolio data. The plot is zoomable and specific data can be shown once the
cursor is placed on the graph.
}
\details{
For all bar plots, green represents profit and red represents loss.
This function also calculates gmv, cumulative P&L and return rates if
these statistics are not in the data frame. Number of contracts will all be zero if the column
is missing.
}

