% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/next_trading_day.R
\name{next_trading_day}
\alias{next_trading_day}
\title{Next trading day}
\usage{
next_trading_day(d)
}
\arguments{
\item{d}{Date relative to which to get the next trading day.}
}
\value{
Date The next trading day of date d.
}
\description{
This function gets the next trading day as defined by is.trading.day()
}

