% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slice_data.R
\name{slice_data}
\alias{slice_data}
\title{Slice Data Set}
\usage{
slice_data(x, input, capital.num)
}
\arguments{
\item{x}{is an input data frame that is to be sliced according to the user's
selection of strategy, portfolio and instrument.}

\item{input}{is a Shiny object that contains all the shiny inputs as a list.}

\item{capital.num}{is the amount of allocated capital for the portfolio,
numeric type. The function will use capital.num to calculate return rates
if available, or use everyday's GMV to calculate return rates instead.}
}
\value{
A small data set that has been retrieved from the previous data set.
}
\description{
This function is part of the Shiny reactive functions. The function looks at
what strategy, portfolio and instrument the user selects, and then slices
the data set into different pieces or summarizes the data set into a summary.
}

