% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat_calculation.R
\name{stat_calculation}
\alias{stat_calculation}
\title{Calculate Statistics}
\usage{
stat_calculation(x.list)
}
\arguments{
\item{x.list}{is the input data set. Such data set has been sliced by
slice_data function, and is labeled as "intermediate data set".}
}
\value{
f A list that contains all calculated numbers.
}
\description{
This helper function calculates the summary of the input data set. The
function takes in the sliced data set and the calculated trading frequency,
and calculate statistics about the market value, returns and volatility for
later display.
}

