% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sum_data.R
\name{sum_data}
\alias{sum_data}
\title{Summarize data for the overall portfolio}
\usage{
sum_data(x, sector.selected = NULL, capital.num = NULL)
}
\arguments{
\item{x}{A data frame that contains data for individual commodities.}

\item{sector.selected}{The sector that the function is doing summarization
for. When "sector.selected = NULL", the function performs summarization across
the whole data set.}

\item{capital.num}{The constant number of allocated capital to calculate return,
number. Default value is NULL. When "capital.num = NULL", the function uses
gross market value to calculate return rates for each row.}
}
\value{
A data frame that summarizes the market values and returns across all
commodities.
}
\description{
This function takes in a data set and returns the summary for each day's
net market value, profit and loss, contract number and return rate.
}

