% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trade_freq.R
\name{trade_freq}
\alias{trade_freq}
\title{Find Trading Frequency}
\usage{
trade_freq(x)
}
\arguments{
\item{x}{The input data set whose trading period we are interested in}
}
\value{
trade.freq The number that indicates the trading frequency.
}
\description{
This helper function looks for the trading frequency of the input data set.
It can identify daily, weekly, monthly or yearly trading frequency in the
data set.
}

